Refactor: 메시지 이름 변경 - #8
