+++ src/main/java/com/sopt/agoda/common/exception/AgodaException.java
@@ -1,6 +1,6 @@
 package com.sopt.agoda.common.exception;
 
-import com.sopt.agoda.common.response.code.FailMessage;
+import com.sopt.agoda.common.response.message.FailMessage;
 
 public class AgodaException extends RuntimeException{
     private final FailMessage failMessage;
