+++ src/main/java/com/sopt/agoda/common/response/BaseResponse.java
@@ -1,8 +1,8 @@
 package com.sopt.agoda.common.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.sopt.agoda.common.response.code.FailMessage;
-import com.sopt.agoda.common.response.code.SuccessMessage;
+import com.sopt.agoda.common.response.message.FailMessage;
+import com.sopt.agoda.common.response.message.SuccessMessage;
 import org.springframework.http.HttpStatus;
