+++ src/main/java/com/sopt/agoda/common/response/message/FailMessage.java
@@ -1,4 +1,4 @@
-package com.sopt.agoda.common.response.code;
+package com.sopt.agoda.common.response.message;
 
 import org.springframework.http.HttpStatus;
 
@@ -7,9 +7,9 @@ public enum FailMessage {
      * 400 Bad Request
      */
     BAD_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
-    INVALID_REQUEST_PARAMETER_VALUE(HttpStatus.BAD_REQUEST, "유효하지 않은 요청 파라미터 값입니다."),
-    INVALID_REQUEST_JSON(HttpStatus.BAD_REQUEST, "잘못된 JSON 형식입니다."),
-    INVALID_REQUEST_HTTP_METHOD(HttpStatus.METHOD_NOT_ALLOWED, "잘못된 http method입니다."),
+    BAD_REQUEST_PARAMETER_VALUE(HttpStatus.BAD_REQUEST, "유효하지 않은 요청 파라미터 값입니다."),
+    BAD_REQUEST_JSON(HttpStatus.BAD_REQUEST, "잘못된 JSON 형식입니다."),
+    BAD_REQUEST_HTTP_METHOD(HttpStatus.METHOD_NOT_ALLOWED, "잘못된 http method입니다."),
 
     /**
      * 401 Unauthorized
@@ -24,8 +24,8 @@ public enum FailMessage {
     /**
      * 404 Not Found
      */
-    ENTITY_NOT_FOUND(HttpStatus.NOT_FOUND, "대상을 찾을 수 없습니다."),
-    API_NOT_FOUND(HttpStatus.NOT_FOUND, "잘못된 API입니다.."),
+    NOT_FOUND_ENTITY(HttpStatus.NOT_FOUND, "대상을 찾을 수 없습니다."),
+    NOT_FOUND_API(HttpStatus.NOT_FOUND, "잘못된 API입니다.."),
 
     /**
      * 405 Method Not Allowed
